requirements.txt
asgiref==3.2.10
awsebcli==3.19.0
botocore==1.17.62
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-tinymce4-lite==1.8.0
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
jsmin==2.2.2
pathspec==0.5.9
Pillow==7.2.0
pypiwin32==223
python-dateutil==2.8.0
python-http-client==3.3.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
sendgrid==6.4.7
six==1.11.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.2.0
